fix: add src to working directory
The imports with "from src..." only worked for unittest, because
unittest was run from the root directory.  This sets the current
directory to the root when running code.  When trying to run the app, a
ModuleNotFoundError occurs, because the app is in the src folder,
making the src folder the current working directory when running code.

This is fixed by assuming the main working directory to be src and
adding the src folder to the system path for tests.

The # nopep8 comments are added to stop import sorting from moving the
work directory addition to after the imports.