fix: enable debug toolbar when not testing
Originally, the debug toolbar was being rendered only during tests.  It
was mistakenly placed there and can not be utilized, since it is only
used when physically viewing webpages.  Tests do not look for any info
within the toolbar.

The toolbar is now only rendered when the app is not in testing mode.
Currently, it is also set to not intercept redirects, since it is
annoying to do the extra click.  This might be changed as needed.