fix test: clear logged in user
When running tests after the successful login test,
flask_login.current_user still has user info saved from that test.  I
am guessing this occurs because the "remember" option is enabled for
flask_login.login_user.  This has caused assertions, that check to see
if there is no currently logged in user, to fail.

This is fixed by calling flask_login.logout_user() for tests (currently
one test) that are able to successfully log in users.  Having the
logout in tearDown() doesn't work, because it throws an out of context
error.