fix: use BigInteger for StreamingOption.expires_on
The streaming_options table had its expires_on column data type as
Integer.  From the Streaming Availability API, the value for expires_on
is a UNIX timestamp that can be as big as 1000 years into the future.
This resulted in a UNIX timestamp value greater than signed Integer.

The column data type for expires_on has been changed to BigInteger to
compensate for this.

Also switched the position of country_code and service_id column
definitions in StreamingOption for organization style.