feat: implement get_streaming_options API
* Added a route to retrieve streaming options.  This has pagination.
This is for asynchronously retrieving lists of streaming options on the
home page or similar webpages.

* Added a get_streaming_options method to StreamingOption, which will
do the core work in retrieving streaming_options records from the
database.  Having this method in StreamingOption helps with
organization.

* Also added a method in StreamingOption to convert an instance into
JSON, which will be used when returning pure streaming option data in
API responses.