feat: implement search by title
Created the app route and basic HTML.

Also in seed.py, consolidated retrieval of RAPID_API_KEY by removing
the retrieval in this file and utilizing the one in app.py.

The test file is named untest..., and should be renamed to test... when
needed.  This is done so that the tests are not easily run
automatically, because the tests make real network calls to the API.