fix: set main working directory to root
Having the "src" folder as the main working directory caused issues
when VSCode trys to find tests.

All directory references have been updated in reference to the true
root.  As a result, all relative imports have to be changed to
absolute, and the root is appended to the system paths when running
app.py and seed.py.

There is still an unsolved issue when trying to discover the route view
tests.  Instead of the tests, a date and something about "INFO" and
sqlalchemy engine are discovered.