feat: implement homepage movie list page change
Added functionality that allows scrolling through a streaming service's
list of movies on the homepage.  This makes a call to the database and
loads the previous or next page of movies.

A JQuery event handler attachment (event listener) is added to the HTML
element, that contains the left and right arrows and list of movies,
instead of adding individual event handlers to the arrow buttons,
because data is needed from the HTML element.  With this way, there is
less DOM traversal.