refactor, fix: put page in data attr in home.js
Refactored code that got country code from cookie at document-ready
section of home.js, so that it looks cleaner.

When building the HTML elements for a streaming service's list of
movies on the homepage, the page number was previously stored using
JQuery's .data(), which does not show the page number on the HTML.  To
improve clarity, the page number has now been changed to be stored in a
"data-page" attribute.