design: create database schema diagram
There is no table for countries, because only the country code will be
used.

The services and movies tables are based off of the related objects in
Streaming Availability API.  Certain columns/attributes are not null,
because that is the case in the API.

In streaming_options, the id is primary key and not movie_id,
country_code, and service_id, because there can be more than one link
for a particular movie, service, and country.

In posters, movie_id, type, and size are primary keys, because there
should be only one image per unique combination of those.