feat: create movie posters methods
Created a method, get_movie_posters, that will retrieve a list of
MoviePosters when given movie_ids, types, and sizes.  This is currently
used for the API side, to allow for asynchronously getting the posters,
but can be used for anything requiring a list of MoviePoster objects.

Create a method, convert_list_to_dict, to convert a list of MoviePoster
objects to a dictionary, so that it can be returned in an HTTP response.

Also added in Enums for movie poster types and sizes, since those have
limited values and will not change.