test chore: rename movie views test
Renamed test from untest_movie_views.py to test_movie_views.py.

It was named untest... because it was making actual external API calls,
and I did not want it to automatically run when running test suites.
It also helped remind me that it made actual external API calls.

It is now named test... because the API GET request calls to the
external API are now mocked.