feat: display movie posters in search results
Display a movie's poster after doing a search for movie titles.

In the test, only the poster's URL path, and not the whole URL, is
asserted in the HTML.  This was done using Python's urllib.parse module
and allows the assertion to work without hardcoding the URL.
Previously, using the whole URL did not work, because in the HTML, the
URL would be parsed and ampersands in the URL would change to something
similar to "&amp;".

MovieDetailsViewTestCase.test_movie_details_page_without_movie_data_in_l
ocal_database is changed for the same reason.