feat: add bookmarking when seeding
Due to Streaming Availability API's request rate limit, it is not
possible to retrieve all of the movies in one day for seeding this
app's database.

To be able to continue on another day, the next cursor (movie) is saved
for a particular country and streaming service.  If all movies have
been retrieved for a country and service, the cursor value will be
"end", and there will be no more API calls to get movies for
country-service.

In streaming_availability_seeder.py, at the bottom end, line
'if __name__ == "__main__":'
is added, so that the actual app is not started for tests.

In streaming_availability_seeder.py, descriptions were also updated.

The file for storing the cursors, streaming_availability_cursors.json,
is git-ignored, because for newly running instances of this app, there
won't be any previous data to continue off of.