refactor: put country code cookie name in variable
On the Python side, the cookie name was named country_code, which is
standard Python naming convention.  On the JavaScript side, it is named
countryCode.  This caused inconsistencies and increased time spent on
maintenance.

The cookie name for the country code is now saved in a variable and
placed at the top of app.py for easier maintenance.