feat: add country dropdown in navbar
Added the dropdown Select HTML element for choosing a country to
display the website for.  The dropdown includes country flags and two
characters for country names.

JavaScript is added to display the previously saved country and to
store a newly selected country.  The webpage is reloaded after
selecting a new country, because that will allow changing the website
text into another language more easily.

A hardcoded list of the countries available from Streaming Availability
is used in navbar.html.  This should be temporary and is done to
quickly get the feature working.  A dynamic solution should be
implemented, where the countries will be retrieved from the database
and sent to the HTML through Flask and Jinja.  This will require
creating another database table to store country code, name, flag, and
displayed short country name.  The dynamic solution might cause
confusion if all of the streaming services, for a country, no longer
have any free streaming options.