feat: retrieve existing data before storing new
Previously when storing the Show JSON object from Streaming
Availability API, only new entries into the database are allowed.  This
prevented being able to update old entries using the Changes JSON
object and route from the API.

Now, when storing, existing movies, posters, and streaming options are
now retrieved from the database.  If they exist, then the old entries
are updated with the new values from the SA API.

The addition of updating old entries is done this way, because it is
the easiest way and didn't require massive code changes.  An
alternative would be to use PostgresQL's "ON CONFLICT DO UPDATE"
feature, but that will require changing how data is added to the
database.

In store_movie_and_streaming_options(), the logging have been moved
down, because if calling the second conversion function results in an
error, the log for adding a movie would be written, even though it was
not added.