feat: update with changes from external API
Implemented a function to help with updating movies, posters, streaming
options, and etc. by calling Streaming Availability API's "changes"
route, storing the received Show object, and returning an indication if
there is more to retrieve and a next "from" timestamp to save the
position on where to start retrieving.  This function only makes one
API call.