feat: allow multiple requests to get updates
Added a function to make multiple requests to update database with new
data from Streaming Availability API, and for all countries.  It will
save the timestamp of the last change from a response, allowing the
next request for updates to start where the last left off.  The
function also has a limit to the total number of requests to make and
is proportional to the max daily limit of the Streaming Availability
API.