fix: make homepage movie arrow buttons larger
On the homepage, the left and right arrow buttons for scrolling movies
for each streaming service were small and only viewable in the middle
of each streaming service's section.  This made it difficult to scroll
when on mobile devices.

The arrow buttons have now been expanded vertically, filling up the
entire height of each streaming service's section.

In home.test.js, the helper function, createTestDivHelper(), has been
updated with the new HTML.  Note that it has not been updated in a very
long time.