chore: remove temp code
The configuration property REMEMBER_COOKIE_DURATION is used for
Flask-Login.  It was set to 60 seconds for testing purposes.

It is now removed.  The default duration to store the cookie is 365
days.

Note that it might not have been working.  Supposedly, Flask's property
PERMANENT_SESSION_LIFETIME needs to be also set.

The Flask Debug Toolbar is commented now.  It is commented instead of
being removed, because it allows for easy enabling if development
continues in the future.

Other temporary code remain due to not having time to improve them.
They should not affect the core functionality of the app.