fix: align navbar items
The select input for choosing countries aligned to the top of its
container.  This was usually not noticeable, but can be if the navbar
was large.

When on small screens, the navbar dropdown was clumping some elements
together.  Some spacing is added now that only affects when screen
sizes are smaller than Bootstrap medium.