fix: get correct target on homepage scroll
In home.js -> handleServiceMoviesPageChange(), the arrowButton was set
to be the event's target, which was the element clicked.  This caused
finding the element with class "bi-arrow-left" to work incorrectly.
When clicking on the arrow image on the list of movies for a streaming
service on the homepage, it would do nothing or set the page to 1.

The arrowButton should have been set to the event's currentTarget,
which is the arrow button's event listener's selector.  This is the
arrow button's container, and not just the arrow image.