feat: change account password requirements
There were no password length or character requirements.  Users can
create accounts with passwords that are 2 characters long or of only
numbers.

The WTForm for account registration is updated to use a regex validator
and pattern to determine if the password contains one uppercase letter,
one lowercase letter, one number, one special character, and be at
least 6 characters long.

The WTForm for account login is updated to check for password length.
It is not necessary to check characters, since that has already been
verified during registration.  The check for password length is added
to help quick fail.

The User model is updated with the new password requirements as well.
There is no check for the database's password column, because passwords
will be encrypted.