feat: allow upsert to database
Added new methods for Movie, MoviePoster, and StreamingOption to allow
bulk inserting or updating to their databases.

A list of each object's data can be passed in, in dictionary form, and
the methods will create new records if the objects didn't exist before,
or update existing objects in the database with the passed-in data.

For StreamingOption, only inserting is allowed, since it would not be
possible to find an old record based on movie ID, country code, and
service ID.  Per those criteria, there can be multiple streaming option
links, such as for different languages.  The id attribute for
StreamingOption is used only for this app, and could not be looked up
based upon updated data from Streaming Availability API.

For the streaming option tests, the existing test case was renamed from
StreamingOptionTestCase for clarity.  Also for that test case, the
Movie and Service tables needed to be cleared out before all tests,
because other test cases might have left data in them.