test refactor: move class variables into setup
In StreamingOptionIntegrationTestsGetStreamingOptions, moved
COUNTRY_CODE into setUpClass, because it was more appropriate.

In the same test case, removed hardcoded SERVICE_01_ID and got the
service ID from the generated Service.  Having the hardcoded service ID
made it harder to maintain.

Also renamed file from test_streaming_option_model.py to
test_streaming_option.py to clarify that it is not only testing the
StreamingOption model, but the methods in StreamingOption as well, and
to be consistent with the other model test files.