fix: remove duplicates on upsert to database
PostgreSQL does not allow bulk insert/update with duplicate records.

To fix this, instead of storing all the movie, poster, and streaming
option dictionaries in lists, they are stored in a dictionaries, with
keys being the unique identifier for an item and values being the
movie, poster, or streaming option dicts.  Keys are strings, because
they are only used to prevent duplicates.  For example, they can
contain movie ID or movie ID, poster type, and poster size.

When passing the movie and etc. data to the upsert functions, the dicts
are reverted back into lists, using the dict values.

Sets are not used to prevent duplicates, because the data is stored in
dictionaries, and dicts are not hashable.