feat: update updater to use new transform methods
Changed functions in updater file to use new transform methods.  Movie,
poster, and streaming option data are now saved in dictionaries, to
prevent duplicates, and are only inserted/updated into the database
after all data gathering is done.

The deleting of a movie's streaming options is now moved into
get_updated_movies_and_streams_from_one_request().  That is the
top-most level it can be placed.

In get_updated_movies_and_streaming_options(), returns after throwing
an exception and hitting the rate limit are now replaced with logic to
exit the loops, so that data, retrieved so far, can be stored in the
database.

Added a new common constant for seeder and updater, for a preferred
rate limit.  This is less than the hard limit set by Streaming
Availability, so that the updating doesn't use up all of the allowed
number of requests.

Created a helper file and functions that deletes a movie's streaming
options for a country and puts movie data in dictionaries according to
the data's unique identifiers.  These were already in the seeder file
and were recreated in the helper file to reduce duplicated code and
lessen the amount of code in the seeder and updater files.  The seeder
functions are not using these functions yet, and they need to be
refactored.

Essentially rewritten the tests for the updater.  Besides updating the
tests accordingly, the lines of code in each test are reorganized to
make tests more readable.