fix: compensate for missing response message
In streaming_availability_updater.py ->
get_updated_movies_and_streams_from_one_request(), for the case where
the response status code is not 200, a StreamingAvailabilityApiError is
raised, with a message from the response or None.

Instead of None, a default message is added for better clarity.

Also changed the description of StreamingAvailabilityApiError.