refactor: move code logic out of movies routes
The searching for movies by title and movie details page routes were
performing too many responsibilities.  They had API calls and data
processing.

To improve readability and maintainability, that code is moved to a
services layer.  Since the moved code is smaller and simpler, unit
tests are now able to be written.

In the two routes in app.py, missing global exception handling is added.

In the two new methods in AppService, the code for rendering templates
is now replaced with code that raises exceptions or returns data.

Tests for the movie views have been updated to mock the requests module
that is now used in app_service.py.  Also, the "reason"
variable/property in the API response is no longer used, and the error
message that is displayed in the HTML webpage is now custom.