fix: add top-level error handling in route
Added catching for all exceptions for login route, which will give
client-friendly flash messages.  Previously, exceptions would cause a
Flask stack trace to be shown.

Wrapped User.authenticate() database call with error handling, so that
the SQLAlchemy session can be rolled back and the error can be logged.

Added testing for this new addition and registration's parallel, as
well as refactoring common code for getting the URL into the setUp().