fix: change search results and navbar css layout
On the search results webpage:

Added vertical margins to the "Search Results" header for better
visibility.

Changed the HTML tag, a movie's year would be in, from "span" to "p"
and added a bottom margin for better visibility.

Changed movie poster images to have rounded tops for small screen
widths and rounded starts for larger widths, for better aesthetics.  A
"@media" addition to the CSS was needed.  The "!important" keyword was
needed to overrule Bootstrap's styling, which uses "!important".  Note
that the specific numbers in the CSS have to match Bootstrap's numbers
for the small breakpoint and border radius.

For the navbar:

Added spacing to the navbar.

Changed the Home button to only be display in light text, because the
previous black text on hover did not look good.