fix: stop display of a movie across pages
On the homepage, the list of movies for streaming services would
sometimes repeat the same movies across different pages.  This issue is
different from the other current issue, where a movie would be
displayed multiple times due to multiple streaming options.  This issue
happens because the paginated results of streaming options are ordered
by rating, where if there are multiple movies with the same rating,
each call to retrieve a new page of streaming options might return a
list of movies that are in a different order.

This issue is fixed by ordering the list of streaming options by movie
rating, then title, and then ID.

Also added a new parameter to set rating for movie generation in
tests/utilities.movie_generator.