chore: output log to console/stderr
When ever an error occurs on the server side, the error is only logged
in a file.  The render website only allows viewing the console of the
app, not the files, for free accounts.  This makes it impossible to
debug.

A new stream handler is added to the logger.  This will have the logs
output to console / stderr too.