fix test: remove old database test data
For StreamingOptionApiTestCase, the Movie and Service tables were not
cleared before adding new test data.  This caused duplicate key
violations from the database for multiple testcase classes when running
all tests.

This is now fixed by firstly deleting the records for tables Movie and
Service in the setUpClass method in StreamingOptionApiTestCase.